Add email regex; move regex patterns
