+++ src/app/model/_model.ts
@@ -10,3 +10,9 @@ export class ModelIndices {
     return ['&_id', 'account']
   }
 }
+
+export const ID_REGEX = /^[a-f\d]{24}$/i
+export const CODE_REGEX = /^\w[\w-]*$/
+export const TITLE_REGEX = /^\w[\w -]*$/
+export const NAME_REGEX = /^[^\s][^\n]*$/
+export const EMAIL_SCHEMA_REGEX = new RegExp("^$|^(([^<>()\\[\\]\\\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$")
